fix: 도메인 관련 제약사항 추가
- messages.properties 수정
- import.sql 수정
- 테스트 수정

+++ src/main/java/com/baemin/nanumchan/domain/ReviewRepository.java
@@ -19,8 +19,8 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
             nativeQuery = true)
     Optional<Double> getAvgRatingByWriterId(Long writerId);
 
-    Long countByWriterId(Long writerId);
+    int countByWriterId(Long writerId);
 
-    Long countByChefId(Long chefId);
+    int countByChefId(Long chefId);
 
 }
